requirements.txt
python-telegram-bot==8.1.1
redis==2.10.6
youtube_dl==2017.10.29
requests==2.11.1
